@@ -11,6 +11,7 @@ class codeList {
     'Buy Airtime/Data From Bank',
     'Borrow Airtime',
     'Call Customer Care',
+    'Link NIN',
     'Know My Phone Number',
     'Transfer Airtime (Share N Sell)',
     'Change Transfer Pin (default pin: 0000)',
@@ -85,18 +86,19 @@ class codeList {
   ];
 
   static final mtnUssdCode = [
-    '*556#',
+    '*310#',
     '*123*1*3#',
-    '*131*4#',
+    '*323#',
     '*559#',
-    '*555*PIN#',
-    '*131#',
+    '*311*PIN#',
+    '*312#',
     '*904#',
-    '*606#',
-    '180',
+    '*303#',
+    '300',
+    '*966#',
     '*123*1*1#',
-    '*600*num*amt*pin#',
-    '*600*OldPin*NewPin*NewPin#',
+    '*321#',
+    '*321*1#',
     '*123*5#',
     '*133*1*num#',
     '*123*1*2#',
@@ -170,7 +172,6 @@ class codeList {
   static final airtelTitles = [
     'Check Airtime Balance',
     'Check Data Balance',
-    'Check Data Balance 2',
     'Know My Phone Number',
     'Cancel data Auto-Renewal',
     'Call Customer Care',
@@ -233,22 +234,21 @@ class codeList {
   ];
 
   static final airtelUssd = [
-    '*123#',
-    '*140#',
-    '*141*712*0#',
+    '*310#',
+    '*323#',
     '*121*3*4#',
     'text STOPAUTORENEW to 440',
-    '111',
+    '300',
     '*121#',
-    '*126*PIN#',
-    '*444#',
+    '*311*PIN#',
+    '*312#',
     '*555*PIN#',
-    '*500#',
+    '*303#',
     '*140*phonenumber#',
     '*791#',
     '*141#',
-    'text 2U RECIPENTNUMBER AMOUNT PIN to 432',
-    'text Pin OLD-PIN NEW-PIN to 432',
+    '*321#',
+    '*321#',
     '*312#',
     '*141*50#',
     '*141*100#',
@@ -300,7 +300,7 @@ class codeList {
   static final nineMobileTitles = [
     'Check Account Balance',
     'Check Data Balance',
-    'Self Service',
+    'Self Service / Buy Airtime & Data',
     'Know My Phone Number',
     'Recharge Airtime',
     'Borrow Airtime',
@@ -340,19 +340,19 @@ class codeList {
   ];
 
   static final ninmobileCodes = [
-    '*232#',
-    '*228#',
-    '*200#',
+    '*310#',
+    '*323#',
+    '*312#',
     '*248#',
-    '*222*PIN#',
-    '*665#',
+    '*311*PIN#',
+    '*303#',
     '*665*2#',
     '*266*1*Phone Number#',
     'text STOP to 229',
     '*229*0#',
-    '200',
-    '*223*pin*amt*num#',
-    '*247*OldPin*NewPin#',
+    '300',
+    '*321#',
+    '*321#',
     '*229*3*8#',
     '*229*3*1#',
     '*229*3*10#',
@@ -385,6 +385,7 @@ class codeList {
     'Check Airtime Balance',
     'Check Data Balance',
     'Recharge Airtime',
+    'Buy Data',
     'Know My Phone Number',
     'Call Customer Care',
     'Borrow Airtime',
@@ -432,18 +433,19 @@ class codeList {
   ];
 
   static final gloUssdCodes = [
-    '#124#',
-    '*127*0#',
-    '*123*PIN#',
+    '*310#',
+    '*323#',
+    '*311*PIN#',
+    '*312#',
     '*135*8#',
-    '121',
-    '*321#',
+    '300',
+    '*303#',
     'text CANCEL to 127',
     '*125*num#',
     '*127*01*num#',
-    '*131*num*amt*pin#',
+    '*321#',
     '*132*OldPin*NewPin*NewPin#',
-    '*777#',
+    '*312#',
     '*127*32#',
     '*127*14#',
     '*127*51#',
